topic methodes en endpoints van forum verplaats naar aparte ForumTopic Service & Controller
