+++ src/main/java/com/autiwomen/auti_women/services/ForumService.java
@@ -178,57 +178,6 @@ public class ForumService {
         return commentedForumDtos;
     }
 
-    public List<Forum> getForumsByTopic(String topic) {
-        List<Forum> forums = forumRepository.findAll();
-        List<Forum> forumsByTopic = new ArrayList<>();
-        for (Forum forum : forums) {
-            if (forum.getTopic().equals(topic)) {
-                forumsByTopic.add(forum);
-            }
-        }
-        return forumsByTopic;
-    }
-
-    public Set<String> getUniqueTopics() {
-        List<Forum> forums = forumRepository.findAll();
-        Set<String> uniqueTopics = new HashSet<>();
-        for (Forum forum : forums) {
-            uniqueTopics.add(forum.getTopic());
-        }
-        return uniqueTopics;
-    }
-
-    public List<String> getSortedUniqueTopics() {
-        List<Forum> forums = forumRepository.findAll();
-        Map<String, Integer> topicFrequency = new HashMap<>();
-        for (Forum forum : forums) {
-            String topic = forum.getTopic();
-            topicFrequency.put(topic, topicFrequency.getOrDefault(topic, 0) + 1);
-        }
-        return topicFrequency.entrySet()
-                .stream()
-                .sorted(Map.Entry.<String, Integer>comparingByValue().reversed())
-                .map(Map.Entry::getKey)
-                .collect(Collectors.toList());
-    }
-
-    public Map<String, Integer> getTopicFrequency() {
-        List<Forum> forums = forumRepository.findAll();
-        Map<String, Integer> topicFrequency = new HashMap<>();
-        for (Forum forum : forums) {
-            String topic = forum.getTopic();
-            if (topicFrequency.containsKey(topic)) {
-                topicFrequency.put(topic, topicFrequency.get(topic) + 1);
-            } else {
-                topicFrequency.put(topic, 1);
-            }
-        }
-        return topicFrequency.entrySet()
-                .stream()
-                .sorted(Map.Entry.<String, Integer>comparingByValue().reversed())
-                .collect(LinkedHashMap::new, (m, e) -> m.put(e.getKey(), e.getValue()), LinkedHashMap::putAll);
-    }
-
     public List<ForumDto> getForumsSortedByLikes() {
         List<Forum> forums = forumRepository.findAll();
         forums.forEach(this::updateForumCounts);
